+++ raspberry-pi/app/motors/__init__.py
@@ -1,6 +1,6 @@
 """Hardware-separated motor interfaces and motion policy."""
 
-from .homing import HomingError, HomingResult, home_stepper
+from .homing import HomingError, HomingResult, home_stepper, home_stepper_for_seconds
 from .stepper import LgpioStepper, StepperError
 
 __all__ = [
@@ -9,4 +9,5 @@ __all__ = [
     "LgpioStepper",
     "StepperError",
     "home_stepper",
+    "home_stepper_for_seconds",
 ]